+++ packages/core/admin/admin/src/components/CommandPalette/Main.js
