+++ packages/core/admin/admin/src/components/CommandPalette/icons/Computer.js
