+++ packages/core/admin/admin/src/components/CommandPalette/index.js
