+++ packages/core/admin/admin/src/components/CommandPalette/pages/ContentManager.js
