+++ packages/core/admin/admin/src/components/CommandPalette/pages/ContentTypeBuilder.js
