+++ packages/core/admin/admin/src/components/CommandPalette/pages/Theme.js
