+++ packages/core/admin/admin/src/pages/Admin/index.js
@@ -13,7 +13,7 @@ import { HTML5Backend } from 'react-dnd-html5-backend';
 
 import GuidedTourModal from '../../components/GuidedTour/Modal';
 import LeftMenu from '../../components/LeftMenu';
-import CmdK from '../../components/CmdK';
+import CommandPalette from '../../components/CommandPalette';
 import AppLayout from '../../layouts/AppLayout';
 import { useMenu, useConfigurations } from '../../hooks';
 import { createRoute } from '../../utils';
@@ -110,7 +110,7 @@ const Admin = () => {
         </Suspense>
         <GuidedTourModal />
 
-        <CmdK />
+        <CommandPalette />
         {showTutorials && <Onboarding />}
       </AppLayout>
     </DndProvider>
